# pulse/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
host: localhost:6532
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/kevin/node-pulse
  description: |-
    Node Pulse is a distributed monitoring system for tracking node health, metrics, and alerts in real-time.

    ## Features
    - Real-time beacon heartbeat monitoring
    - Multi-metric data collection and analysis
    - Configurable alert rules with suppression
    - Webhook notifications for alerts
    - Historical data queries and export
    - Role-based access control (RBAC)

    ## Authentication
    Most endpoints require authentication. Use the `/api/v1/auth/login` endpoint to obtain a session token.
    Include the token in the `Authorization` header: `Bearer <your-token>`
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Node Pulse API
  version: "1.0"
paths:
...
  /api/config:
    get:
      consumes:
      - application/json
      description: Get current application configuration (admin only, passwords redacted)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current configuration
      tags:
      - config
  /api/config/validate:
    get:
      consumes:
      - application/json
      description: Validate current application configuration (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Validate configuration
      tags:
      - config
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"